+++ examples/jetstream/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lex-gql-jetstream-example
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - lex-gql@0.4.0
+
 ## 1.0.6
 
 ### Patch Changes