+++ examples/relay/CHANGELOG.md
@@ -1,5 +1,12 @@
 # relay-example
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - lex-gql@0.4.0
+
 ## 0.0.3
 
 ### Patch Changes